requirements.txt
fastapi==0.95.0
pydantic[email]~=1.10.7
uvicorn==0.21.1
uvloop==0.17.0

psycopg2-binary==2.9.5
sqlalchemy==2.0.7
alembic==1.10.2
asyncpg==0.27.0

redis==4.5.3

PyJWT==2.6.0
python-jose==3.3.0
cryptography==40.0.1

python-dotenv==1.0.0